+++ scripts/start_db.sh
@@ -11,4 +11,4 @@ sleep 3
 
 # Run a migration to create your database tables with Prisma Migrate
 # && Populate database with sample data
-npx prisma migrate dev --name init && npx ts-node populate_db.ts
+npx prisma migrate dev --name init && npx ts-node scripts/populate_db.ts
