+++ src/server.ts
@@ -1,0 +1,5 @@
+import express from "express";
+
+const app = express();
+
+export default app;
