+++ vercel.json
@@ -2,14 +2,14 @@
   "version": 2,
   "builds": [
     {
-      "src": "index.ts",
+      "src": "src/index.ts",
       "use": "@now/node"
     }
   ],
   "rewrites": [
     {
       "source": "/(.*)",
-      "destination": "index.ts"
+      "destination": "src/index.ts"
     }
   ],
   "git": {
